INTERLOK-633 Added some null protection INTERLOK-633 #time 1h
JuanR testing against salesforce, turns out the entity from the response
can be null (it wasn't clear from the javadocs). Also added in a null
check for the response headers just in case they're null as well (they
shouldn't be, because we should always have headers, even if it's a
Date: ).

Made logger transient so that we don't break XStream.