INTERLOK-2658 Exclude DMI_BLOCKING_METHODS_ON_URL
Turns out that uri.equals(getRquestURL().toURI()  and
uri.toURL().equals(getRequestingURL()) cannot be considered equivalent.
Therefore to not change behaviour we turn off the spotbugs check for that
error.

Effectively, while the code may be considered "non-performant"; we're in
an HTTP session, so we have to be doing DNS lookups, so it's largely
inconsequential.